+++ crates/web-client/src/models/mod.rs
@@ -46,6 +46,7 @@ pub mod account_type;
 pub mod address;
 pub mod advice_inputs;
 pub mod advice_map;
+pub mod asset_callback_flag;
 pub mod asset_vault;
 pub mod auth;
 pub mod auth_scheme;